+++ sample/src/desktopMain/kotlin/com/github/panpf/sketch/sample/ui/components/zoomimage/core/BufferedBitmapImageValue.kt
@@ -2,7 +2,7 @@ package com.github.panpf.sketch.sample.ui.components.zoomimage.core
 
 import com.github.panpf.sketch.cache.MemoryCache
 
-class BufferedImageValue(
+class BufferedBitmapImageValue(
     override val image: BufferedBitmapImage,
     override val extras: Map<String, Any?> = emptyMap()
 ) : MemoryCache.Value {